+++ src/user/user.module.ts
@@ -1,5 +1,5 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { UserController } from './user.controller';
+import { UserController } from './controllers/user.controller';
 import { UserService } from './services/user.service';
 import { RegisterService } from './services/register.service';
 import { UserSchema } from './models/user.model';
